update for exit bit norm removed staic stuff added prompt flag
