+++ .vscode/c_cpp_properties.json
@@ -1,0 +1,18 @@
+{
+	"configurations": [
+	  {
+		"name": "macos-clang-x64",
+		"includePath": [
+		  "${workspaceFolder}/**"
+		],
+		"compilerPath": "/usr/bin/clang",
+		"cStandard": "${default}",
+		"cppStandard": "${default}",
+		"intelliSenseMode": "macos-clang-x64",
+		"compilerArgs": [
+		  ""
+		]
+	  }
+	],
+	"version": 4
+  }
