+++ .vscode/launch.json
@@ -1,0 +1,13 @@
+{
+	"version": "0.2.0",
+	"configurations": [
+	  {
+		"name": "C/C++ Runner: Debug Session",
+		"type": "lldb",
+		"request": "launch",
+		"args": [],
+		"cwd": "/Users/chuhlig/Desktop/merged_minishell/",
+		"program": "/Users/chuhlig/Desktop/merged_minishell/minishell"
+	  }
+	]
+  }
