+++ include/env.h
@@ -6,7 +6,7 @@
 /*   By: chuhlig <chuhlig@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/08 16:53:39 by dkaiser           #+#    #+#             */
-/*   Updated: 2024/12/24 16:21:50 by chuhlig          ###   ########.fr       */
+/*   Updated: 2025/01/14 16:50:55 by chuhlig          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,7 @@ typedef struct s_env
 	struct s_env	*next;
 }					t_env;
 
+void	free_env_node(t_env *node);
 void	getenvlst(t_env **env, char **en);
 void	free_envlst(t_env **env);
 char	*env_get(t_env *env, char *name);
@@ -34,5 +35,7 @@ int		echo(char **av);
 int		pwd(t_env *env);
 int		cd(t_env **env, char **args);
 int		ft_env(t_env *env);
+int		builtin_exit(char **args, t_env **env);
+t_env	*env_new(char *name);
 
 #endif
