+++ lib/libft/ft_strcat.c
@@ -6,7 +6,7 @@
 /*   By: chuhlig <chuhlig@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/09 13:06:44 by chuhlig           #+#    #+#             */
-/*   Updated: 2025/01/09 13:07:15 by chuhlig          ###   ########.fr       */
+/*   Updated: 2025/01/14 14:09:49 by chuhlig          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
