+++ lib/libft/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: chuhlig <chuhlig@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/18 19:03:14 by chuhlig           #+#    #+#             */
-/*   Updated: 2024/12/18 19:05:01 by chuhlig          ###   ########.fr       */
+/*   Updated: 2025/01/14 14:09:59 by chuhlig          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
