+++ lib/libft/ft_strcpy.c
@@ -6,7 +6,7 @@
 /*   By: chuhlig <chuhlig@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/09 14:38:30 by chuhlig           #+#    #+#             */
-/*   Updated: 2025/01/09 14:38:53 by chuhlig          ###   ########.fr       */
+/*   Updated: 2025/01/14 14:10:06 by chuhlig          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
