+++ src/env_to_strlst.c
@@ -6,7 +6,7 @@
 /*   By: chuhlig <chuhlig@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/17 19:22:28 by chuhlig           #+#    #+#             */
-/*   Updated: 2024/12/17 19:22:36 by chuhlig          ###   ########.fr       */
+/*   Updated: 2025/01/14 14:10:49 by chuhlig          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
