+++ src/init.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   init.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dkaiser <dkaiser@student.42heilbronn.de    +#+  +:+       +#+        */
+/*   By: chuhlig <chuhlig@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/24 15:02:24 by dkaiser           #+#    #+#             */
-/*   Updated: 2024/06/24 15:25:57 by dkaiser          ###   ########.fr       */
+/*   Updated: 2025/01/13 17:45:46 by chuhlig          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
