+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: chuhlig <chuhlig@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/22 17:14:03 by dkaiser           #+#    #+#             */
-/*   Updated: 2024/12/17 19:26:42 by chuhlig          ###   ########.fr       */
+/*   Updated: 2025/01/14 15:29:20 by chuhlig          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,17 @@
 
 int	main(int argc, char *argv[], char *envp[])
 {
-	t_env	*env;
+	t_env		*env;
+	static int	promptflag;
 
 	env = NULL;
+	promptflag = 0;
 	if (!argc && !argv)
 		return (1);
 	if (init())
 		return (1);
 	getenvlst(&env, envp);
-	repl("Minishell $ ", &env);
+	repl("Minishell $ ", &env, &promptflag);
+	free_envlst(&env);
+	return (0);
 }
