+++ src/signal_handling.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   signal_handling.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dkaiser <dkaiser@student.42heilbronn.de    +#+  +:+       +#+        */
+/*   By: chuhlig <chuhlig@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/24 15:08:43 by dkaiser           #+#    #+#             */
-/*   Updated: 2024/06/25 13:33:26 by dkaiser          ###   ########.fr       */
+/*   Updated: 2025/01/14 14:11:29 by chuhlig          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
